OpenAITelegramBOT/requirements.txt
python-telegram-bot==13.12
python-dotenv==0.20.0
openai
redis